Changes in registering/SignIn
