+++ app-android-xamarin-Doctor-side/Moodis/Moodis/Feature/Register/RegisterActivity.cs
@@ -1,14 +1,10 @@
 ﻿using Android.App;
-using Android.Content;
 using Android.OS;
-using Android.Runtime;
-using Android.Util;
 using Android.Views;
 using Android.Views.InputMethods;
 using Android.Widget;
 using Moodis.Constants.Enums;
 using Moodis.Extensions;
-using System.Reflection;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
@@ -140,21 +136,6 @@ namespace Moodis.Feature.Register
             else
             {
                 Toast.MakeText(this, Resource.String.stronger_password, ToastLength.Short).Show();
-            }
-            progressBar.Visibility = ViewStates.Gone;
-            registerButton.Enabled = true;
-        }
-
-        private async Task DeleteUser()
-        {
-            progressBar.Visibility = ViewStates.Visible;
-            progressBar.BringToFront();
-            registerButton.Enabled = false;
-
-            var response = await registerViewModel.DeleteUser();
-            if (response != Response.OK)
-            { 
-                Log.Error(Class.Name, MethodBase.GetCurrentMethod().Name + ": " + response.ToString());
             }
             progressBar.Visibility = ViewStates.Gone;
             registerButton.Enabled = true;
