+++ app-android-xamarin-Doctor-side/Moodis/Moodis/Moodis.csproj
@@ -11,7 +11,7 @@
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>Moodis</RootNamespace>
-    <AssemblyName>MoodisDoctors</AssemblyName>
+    <AssemblyName>Moodis</AssemblyName>
     <FileAlignment>512</FileAlignment>
     <AndroidApplication>True</AndroidApplication>
     <AndroidResgenFile>Resources\Resource.designer.cs</AndroidResgenFile>
